refactor(dashboard): move TimeRange to util/chart and format helpers to util/utils
TimeRange was exported from dashboard-financial-history.screen and
imported by 5 chart components and 2 screens. Move it to src/util/chart.ts
so chart components no longer depend on a screen module.

Also move formatChf / formatChfOrDash from summary-card.tsx to util/utils.ts
to keep the component file free of format helpers.
